Refactor: ConceptProcessing: Split ConceptProcessing into different task processing classes for each task type.
